Remove the Newline token type from the lexer; add the Comment type.
Newline runes are now treated as whitespace. Comments are their own
type, instead of being treated as whitespace. The IsNewline method
tests if a token should be treated as a newline. This includes
whitespace and comments containing \n, or line comments (which may not
contain \n if they occur immediately before EOF).