Remove TypeName nodes, use Selector and Identifier instead.
This also exposed the ambiguity with composite literals inside of the
initialization of if, for, and switch statements. After some despair,
I took a look at how it was done by the Go standard library parser.
They coutn expression nesting levels and disallow composite literals
on negative levels. Then, if, switch, and for set the level to -1
around their initial statements. Now we do this too!

Fixes #3