Remove the ParemeterList AST node.
This greatly simplifies the representation of functions, methods, and
anything using a Signature. The cost is that we lose a bit of token
location information. Specifically, before this change, we had the
location of the '(' and ')' tokens for any parameter list. Now, we
only have the location of the start and end of a signature. This seems
like a good trade; I don't see the location information of Signatures
being too useful.