Remove errors.Add.
This method was cute, but made the code harder to read and harder to
test. Harder to read, because it looks like it unconditionally added
the error, but it only added it if non-nil. Harder to test, because it
hid err != nil branches that could be flagged as untested by a
coverage tool.