Don't pass needless arguments to Declaration Checks.
ConstSpec and constSpecView already have access to Iota. They don't
need it again. Also, Declarations don't need their symbol table,
because they already have it as a field. This was necessary for
package-level declarations that needed their file-level symbol table,
which would not be passed via a recursive call from a declaration in a
separate file. For Declarations that appear in blocks,
DeclarationStatement will need to set their syms field. Not a problem.